software/decryptor/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pynacl>=1.5.0
python-multipart>=0.0.6